+++ Assets/Resources/config.json
@@ -12,5 +12,5 @@
   "canvasY": 6,
   "canvasZ": 100,
   "feedbackSize": 75,
-  "feedbackColor": "red"
+  "feedbackColor": "black"
 }
