+++ Assets/Scripts/ManageTrials.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12f142919a225484f91722405ec39851
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
