+++ Assets/Scripts/ReadConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 453e06e8eee1c2e46846a7e8cfd275a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
