+++ Assets/Scripts/RunExperiment.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.IO;
 using UnityEngine;
 
 public class RunExperiment : MonoBehaviour {
@@ -13,12 +12,12 @@ public class RunExperiment : MonoBehaviour {
     public Transform cameraManager;
 
     // The Config options
-    private Config config;                  // The configuration file specifying certain experiment-wide parameters
+    private ReadConfig.Config config;       // The configuration file specifying certain experiment-wide parameters
     private string inputFile;               // A JSON file holding the information for every trial to be run
     private bool targetCamera;              // A boolean to determine if the object follows the user's head or not
 
     // Experiment-Dependent Variables
-    private Trial[] trials;                 // The input file converted to an array of Trial objects 
+    private ManageTrials.Trial[] trials;    // The input file converted to an array of Trial objects 
     private IEnumerator movementCoroutine;  // The coroutine responsible for moving the object in the world
     private Transform headPos;              // The location of the camera rig relevant to the scene
 
@@ -27,8 +26,6 @@ public class RunExperiment : MonoBehaviour {
     private bool isRunning;                 // Tracks whether or not a trial is currently active
     private Vector3 targetPos;              // The target that the moving object aims for
     private float trialStart;               // Track the time that the current trial began
-    private bool waiting;                   // Boolean to track whether we're waiting between trials
-    private float waitTime;                 // Timer to track how long we've been waiting
     private string objName;                 // The name of the prefab object used for the given trial
     private Vector3 startPos;               // The initial position of the moving object for the current trial
     private Transform obj;                  // The prefab object that will be instantiated
@@ -39,91 +36,50 @@ public class RunExperiment : MonoBehaviour {
     private string posString;
     private float hideTime;
 
-    [System.Serializable]
-    public class Config
-    {
-        public int subjNum;
-        public int subjSex;
-        public string dataFile;
-        public bool targetCamera;
-        public bool trackHeadPos;
-        public float initCameraX;
-        public float initCameraY;
-        public float initCameraZ;
-        public bool showFeedback;
-        public float canvasX;
-        public float canvasY;
-        public float canvasZ;
-        public int feedbackSize;
-        public string feedbackColor;
-    }
-
-    [System.Serializable]
-    class TrialArray
-    {
-        // this wrapper class is a workaround for how Unity's JsonUtility
-        // handles top-level JSON objects
-        public Trial[] trials;
-    }
-
-    [System.Serializable]
-    public class Trial
-    {
-        public int trialNum;            // the trial number
-        public float startDist;         // the starting distance of the object (in meters)
-        public float velocity;          // the speed the object is moving (in meters / second)
-        public float timeVisible;       // the amount of time this object should be visible before disappearing
-        public string objType;          // the names of the potential object prefabs that can be instantiated
-        public float objScaleX;
-        public float objScaleY;
-        public float objScaleZ;
-        public float rotationSpeed;     // the speed at which the object should rotate each frame
-    }
-
-    public Config LoadConfig(string configFile)
-    {
-        try
-        {
-            string filepath = configFile.Replace(".json", "");
-            TextAsset jsonString = Resources.Load<TextAsset>(filepath);
-            Config config = JsonUtility.FromJson<Config>(jsonString.ToString());
-            return config;
-        }
-        catch (System.Exception e)
-        {
-            print("Exception: " + e.Message);
-            return null;
-        }
-    }
-
 
     /**
-     * Given a path to a JSON file containing the parameters for each
-     * trial in the experiment, creates a Trial object with the correct
-     * parameters for each entry in the input file.
+     * Initializes all trial data once the experiment begins. This includes loading the
+     * config file, loading in the trial-information, and setting all experiment-wide variables.
      */
-     public Trial[] LoadTrialData(string path, float time)
+    void Start()
     {
-        try
-        {
-            StreamReader sr = new StreamReader(path);
-            string jsonString = sr.ReadToEnd();
-            TrialArray trialData = JsonUtility.FromJson<TrialArray>(jsonString);
+        // Set the target framerate for the application
+        Application.targetFrameRate = 75;
 
-            // Initialize the TrialData array to be the correct size in 
-            // the experiment's data manager
-            dataManager.InitDataArray(trialData.trials.Length, time);
-            return trialData.trials;
-        }
-        catch (System.Exception e)
-        {
-            Debug.Log("Exception: " + e.Message);
-            return null;
-        }
+        // Load the config file
+        config = GetComponent<ReadConfig>().LoadConfig("config.json");
+
+        // Set the configurations based on the config file
+        targetCamera = config.targetCamera;
+        uiManager.SetFeedbackColor(config.feedbackColor);
+        uiManager.SetCanvasPosition(config.canvasX, config.canvasY, config.canvasZ);
+        uiManager.SetFeedbackSize(config.feedbackSize);
+
+        // Load the data from the desired input file
+        trials = GetComponent<ManageTrials>().LoadTrialData(config.dataFile, Time.time);
+        //this.trials = LoadTrialData(inputFile, Time.time);
+
+        // Initialize the TrialData array to be the correct size in 
+        // the experiment's data manager
+        dataManager.InitDataArray(trials.Length, Time.time);
+
+        // Add the config info to the data manager
+        dataManager.SetConfigInfo(config.subjNum, config.subjSex, config.dataFile, config.showFeedback, config.feedbackColor, config.targetCamera, config.trackHeadPos);
+
+        // Initialize global variables
+        curTrial = 0;
+        isRunning = false;
+
+        // Set the initial position of the participant
+        cameraManager.position = new Vector3(config.initCameraX, config.initCameraY, config.initCameraZ);
+
+        // Set the head position transform to track the participant's movements
+        headPos = GameObject.Find("Camera (eye)").transform;
     }
 
+
     /**
-     * Initialize the given trial.
+     * Initialize the current trial.
      */
     public void InitializeTrial()
     {
@@ -132,16 +88,15 @@ public class RunExperiment : MonoBehaviour {
         {
             Debug.Log("Initializing trial " + curTrial);
 
+            // Stop displaying the feedback text
             uiManager.ResetFeedbackMsg();
 
             // Set the target for this trial (will be updated each iteration of Update function if targetCamera is true)
             targetPos = new Vector3(viveCamera.position.x, viveCamera.position.y, viveCamera.position.z);
-
-
             Debug.Log("TARGET POSITION: " + viveCamera.position.x + " " + viveCamera.position.y + " " + viveCamera.position.z);
 
             // Get the current trial from the data array
-            Trial trial = trials[curTrial];
+            ManageTrials.Trial trial = trials[curTrial];
 
             // Set the target position for the moving object
             if (targetCamera)
@@ -237,10 +192,6 @@ public class RunExperiment : MonoBehaviour {
         dataManager.AddTrial(curTrial, trials[curTrial - 1].startDist, startPos, trials[curTrial - 1].velocity, trials[curTrial - 1].rotationSpeed,
             trials[curTrial - 1].timeVisible, objName, trialStart, trialEnd, receivedResponse, respTime);
 
-        // Wait to start the next trial
-        waiting = true;
-        waitTime = 0.0f;
-
         if (config.trackHeadPos) dataManager.WritePosData();
     }
 
@@ -253,47 +204,6 @@ public class RunExperiment : MonoBehaviour {
     public bool CheckTrialRunning()
     {
         return isRunning;
-    }
-
-
-    /**
-     * Initializes all trial data once the experiment begins. Starts
-     * tracking the time that events are occurring.
-     */
-    
-    void Start()
-    {
-        // Set the target framerate for the application
-        Application.targetFrameRate = 75;
-
-        // Load the config file
-        this.config = LoadConfig("config.json");
-
-        // Set the configurations based on the config file
-        inputFile = config.dataFile;
-        targetCamera = config.targetCamera;
-        uiManager.SetFeedbackColor(config.feedbackColor);
-        uiManager.SetCanvasPosition(config.canvasX, config.canvasY, config.canvasZ);
-        uiManager.SetFeedbackSize(config.feedbackSize);
-
-        // Add the config info to the data manager
-        dataManager.SetConfigInfo(config.subjNum, config.subjSex, config.dataFile, config.showFeedback, config.feedbackColor, config.targetCamera, config.trackHeadPos);
-
-        // Load the data from the desired input file
-        this.trials = LoadTrialData(inputFile, Time.time);
-
-        // Initialize global variables
-        curTrial = 0;
-        waiting = true;
-        waitTime = 0.0f;
-        isRunning = false;
-
-        // Set the initial position of the participant
-        cameraManager.position = new Vector3(config.initCameraX, config.initCameraY, config.initCameraZ);
-
-        // Set the head position transform to track the participant's movements
-        headPos = GameObject.Find("Camera (eye)").transform;
-
     }
